+++ NEWS.md
@@ -1,3 +1,13 @@
+# mirai 0.8.2
+
+* `dispatcher()` re-implemented using an innovative non-polling design. Efficient process consumes zero processor usage when idle and features significantly higher throughput and lower latency.
+  + Arguments 'pollfreqh' and 'pollfreql' removed as no longer applicable.
+* Server and dispatcher processes exit automatically if the connection with the client is dropped. This significantly reduces the likelihood of orphaned processes.
+* `launch()` exported as a utility for easily re-launching daemons that have timed out, for instance.
+* Correct passthrough of `...` variables in the `daemons()` call.
+* Requires nanonext >= 0.8.1.
+* Internal performance enhancements.
+
 # mirai 0.8.1
 
 * Fixes issue where daemon processes may not launch for certain setups (only affecting binary package builds).